MoE a16w-mix: extract shared helpers to utils.py + modernize ptr/waitcnt (ISA byte-exact)
Cleanup of the a16w4/a16wi4 package leaf helpers (no codegen change; gemm1+gemm2
AMDGCN verified byte-exact A/B via FLYDSL_DUMP_IR):

- Raw ptr casts -> flydsl llvm-ptr API: _lds_ptr3/_global_base_ptr1 now go through
  a single _int_to_llvm_ptr() using fx.to_llvm_ptr(fx.inttoptr(..., AS.Shared/Global)),
  dropping the hardcoded ir.Type.parse("!llvm.ptr<3|1>") (backend resolves the AS).
- Drop the hand-encoded CDNA3 s_waitcnt shim: flydsl 0.3.0 has the arch-dispatched
  keyword form fx.rocdl.s_waitcnt(lgkmcnt=0); same encoding (ISA identical).
- b_cache_mod: use the CSV-driven b_nt directly (constant default when unset); remove
  the dead/inconsistent per-token U-shape heuristic (moe_kernels always passes b_nt).
- Extract the 18 shared leaf helpers gemm2 imported from gemm1 into a package-local
  utils.py; gemm1 shrinks ~185 LOC. gemm2/__init__/moe_kernels repoint to .utils.

Validated: a16w4 SiTUv2 pytest 6/6 cold (cache cleared), pyflakes + black clean.
